YouTube caption retriever and more...
Added functionality that retrieves the audio transcript (if there is one) at the YouTube video specified by a given URL. Added a module containing a method that installs the necessary third-party packages for the user. This method must be invoked upon starting up the application. The accompanying .txt file must be present for the package installer to function.